ACAN_ESP32: Add ACAN_ESP32::end() to stop the controller and remove ISR
Some applications require stopping the controller (at least temporarily)
for example when starting the OTA process as, during that, getting any
interrupt from the CAN controller may be not desirable, as that could be
increasing the likeliness of an OTA failure (and noone wants to end up
with a corrupt flash).

The OTA case is just one of the possible examples, anyway.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@somainline.org>